add logic for verify kiam action
