+++ cmd/cl001/zz_generated.command.go
@@ -9,7 +9,6 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/giantswarm/awscnfm/v12/cmd/cl001/ac000"
-	"github.com/giantswarm/awscnfm/v12/cmd/cl001/ac010"
 	"github.com/giantswarm/awscnfm/v12/cmd/cl001/ac011"
 	"github.com/giantswarm/awscnfm/v12/cmd/cl001/ac013"
 )
@@ -47,20 +46,6 @@ func New(config Config) (*cobra.Command, error) {
 		}
 
 		ac000Cmd, err = ac000.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
-	var ac010Cmd *cobra.Command
-	{
-		c := ac010.Config{
-			Logger: config.Logger,
-			Stderr: config.Stderr,
-			Stdout: config.Stdout,
-		}
-
-		ac010Cmd, err = ac010.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -113,7 +98,6 @@ func New(config Config) (*cobra.Command, error) {
 	f.Init(c)
 
 	c.AddCommand(ac000Cmd)
-	c.AddCommand(ac010Cmd)
 	c.AddCommand(ac011Cmd)
 	c.AddCommand(ac013Cmd)
