--- pkg/action/cl001/ac010/explainer.go
@@ -1,9 +1,0 @@
-package ac010
-
-import (
-	"context"
-)
-
-func (e *Explainer) explain(ctx context.Context) (string, error) {
-	return "", nil
-}
