--- pkg/action/cl001/ac010/zz_generated.error.go
@@ -1,14 +1,0 @@
-package ac010
-
-import (
-	"github.com/giantswarm/microerror"
-)
-
-var invalidConfigError = &microerror.Error{
-	Kind: "invalidConfigError",
-}
-
-// IsInvalidConfig asserts invalidConfigError.
-func IsInvalidConfig(err error) bool {
-	return microerror.Cause(err) == invalidConfigError
-}
