--- pkg/action/cl004/ac011/explainer.go
@@ -1,9 +1,0 @@
-package ac011
-
-import (
-	"context"
-)
-
-func (e *Explainer) explain(ctx context.Context) (string, error) {
-	return "", nil
-}
